+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- `AccessibilityContainer` now omits accessibility elements where `.accessibilityElementsHidden` is `true`.
+
 ### Added
 
 - `Image` now provides an override to prevent VoiceOver from generating accessibility descriptions.